Update activity folder and files inside
